Fixes and changes on `mccmdhl2.tkapp`
Control keys like Shift and Ctrl does not hide popup window now.

Fix `ValueError: Invalid line & column` when inserting or deleting
text on lines that cursor is not at. (e.g. when trying to accept
an auto-completing suggestion while cursor is not on the line that
suggestions are generated for)

Fix `RecursionError` when sending a lot of keyboard events in a short
moment.

Fix popup flickering on Android (Python 3.9.7 built with Pydroid).
This is because some keys like "!" generate Shift keyboard event and
hide the popup window.